refactor gormx pkg
