+++ internal/oidcx/client_repository.go
@@ -7,7 +7,6 @@ import (
 	"github.com/c1emon/lemontree/pkg/errorx"
 	"github.com/c1emon/lemontree/pkg/gormx"
 	"github.com/c1emon/lemontree/pkg/httpx"
-	"github.com/c1emon/lemontree/pkg/persister"
 	"github.com/pkg/errors"
 	"gorm.io/gorm"
 )
@@ -26,7 +25,7 @@ type GormClientRepository struct {
 
 func NewGormClientRepository() *GormClientRepository {
 	r := &GormClientRepository{
-		db: persister.GetDB(),
+		db: gormx.GetDB(),
 	}
 	_ = r.InitDB()
 	return r
