--- pkg/persister/base.go
@@ -1,5 +1,0 @@
-package persister
-
-type InitDB interface {
-	Init()
-}
